pyproject.toml
[project]
name = "usipipo-landing"
version = "0.1.0"
description = "Landing Page oficial de uSipipo VPN"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [{name = "uSipipo Team", email = "dev@usipipo.com"}]

dependencies = [
    "flask>=3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
